Add server auto message forwarding option
In projects that give clients authority over certain mechanics, informing other clients of changes (like a player moving) often results in writing a lot of repetitive code. As the server mainly just acts as a relay, incoming messages tend to be handled only for the server to immediately forward/distribute them to all other connected clients.

Being able to set a parameter when creating a message to automate this relay behaviour helps simplify code and logic flow. However, to avoid allowing hacked clients to take advantage of this in projects where server-authority is important, this needs to be a setting that is disabled by default.