Make message byte arrays publicly inaccessible
The Message class's Bytes property was only ever publicly accessible to allow various transports to modify a message's data directly, but that can lead to users modifying the array when they probably shouldn't be.

This change moves the access to the Transports namespace, which makes modifying the array directly require a more explicit action on the developer's part.