Remove isBigArray option for Add/Get methods
Whether or not the array is big enough to require two bytes to store its length can be stored in the first bit of the first length byte. This allows arrays with up to 127 elements to automatically use only a single byte to transmit their length, while anything larger (maximum supported length is 32,767) will use two bytes.

This does away with the need to manually set the isBigArray parameter to true in the Add/Get methods when sending large arrays, which was clunky and error-prone.