Only add child links if ToC foldout had a link
Foldouts in the table of contents may not link to anything themselves, especially if they only exist for organizational purposes (like the Methods, Properties, Events, etc. foldouts, as well as foldouts for groups of articles). If the foldout element doesn't have a link, no child element should be created for it.

However, a child element should be created for foldouts which do link to something (such as namespace or class pages) so that the foldout itself doesn't have to be the link.

This change makes table of contents generation account for both of these scenarios.